multilinear mod
